Moved carousel to custom section
